+++ lib/money/money/formatting.rb
@@ -7,7 +7,7 @@ class Money
     #
     # @return [String]
     #
-    # @option *rules [Boolean, String] :display_free (false) Whether a zero
+    # @option rules [Boolean, String] :display_free (false) Whether a zero
     #  amount of money should be formatted of "free" or as the supplied string.
     #
     # @example
@@ -15,7 +15,7 @@ class Money
     #   Money.us_dollar(0).format(:display_free => "gratis") #=> "gratis"
     #   Money.us_dollar(0).format                            #=> "$0.00"
     #
-    # @option *rules [Boolean] :with_currency (false) Whether the currency name
+    # @option rules [Boolean] :with_currency (false) Whether the currency name
     #  should be appended to the result string.
     #
     # @example
@@ -23,28 +23,28 @@ class Money
     #   Money.ca_dollar(100).format(:with_currency => true) #=> "$1.00 CAD"
     #   Money.us_dollar(85).format(:with_currency => true)  #=> "$0.85 USD"
     #
-    # @option *rules [Boolean] :rounded_infinite_precision (false) Whether the
-    #  amount of money should be rounded when using infinite_precision
+    # @option rules [Boolean] :rounded_infinite_precision (false) Whether the
+    #  amount of money should be rounded when using {infinite_precision}
     #
     # @example
     #   Money.us_dollar(100.1).format #=> "$1.001"
     #   Money.us_dollar(100.1).format(:rounded_infinite_precision => true) #=> "$1"
     #   Money.us_dollar(100.9).format(:rounded_infinite_precision => true) #=> "$1.01"
     #
-    # @option *rules [Boolean] :no_cents (false) Whether cents should be omitted.
+    # @option rules [Boolean] :no_cents (false) Whether cents should be omitted.
     #
     # @example
     #   Money.ca_dollar(100).format(:no_cents => true) #=> "$1"
     #   Money.ca_dollar(599).format(:no_cents => true) #=> "$5"
     #
-    # @option *rules [Boolean] :no_cents_if_whole (false) Whether cents should be
+    # @option rules [Boolean] :no_cents_if_whole (false) Whether cents should be
     #  omitted if the cent value is zero
     #
     # @example
     #   Money.ca_dollar(10000).format(:no_cents_if_whole => true) #=> "$100"
     #   Money.ca_dollar(10034).format(:no_cents_if_whole => true) #=> "$100.34"
     #
-    # @option *rules [Boolean, String, nil] :symbol (true) Whether a money symbol
+    # @option rules [Boolean, String, nil] :symbol (true) Whether a money symbol
     #  should be prepended to the result string. The default is true. This method
     #  attempts to pick a symbol that's suitable for the given currency.
     #
@@ -75,10 +75,10 @@ class Money
     #   Money.new(10000000, "INR").format(:south_asian_number_formatting => true) #=> "1,00,000.00"
     #   Money.new(10000000).format(:south_asian_number_formatting => true) #=> "$1,00,000.00"
     #
-    # @option *rules [Boolean, nil] :symbol_before_without_space (true) Whether
-    # a space between the money symbol and the amount should be inserted when
-    # +:symbol_position+ is +:before+. The default is true (meaning no space). Ignored
-    # if +:symbol+ is false or +:symbol_position+ is not +:before+.
+    # @option rules [Boolean, nil] :symbol_before_without_space (true) Whether
+    #   a space between the money symbol and the amount should be inserted when
+    #   +:symbol_position+ is +:before+. The default is true (meaning no space). Ignored
+    #   if +:symbol+ is false or +:symbol_position+ is not +:before+.
     #
     # @example
     #   # Default is to not insert a space.
@@ -90,10 +90,10 @@ class Money
     #   # If set to false, will insert a space.
     #   Money.new(100, "USD").format(:symbol_before_without_space => false) #=> "$ 1.00"
     #
-    # @option *rules [Boolean, nil] :symbol_after_without_space (false) Whether
-    # a space between the the amount and the money symbol should be inserted when
-    # +:symbol_position+ is +:after+. The default is false (meaning space). Ignored
-    # if +:symbol+ is false or +:symbol_position+ is not +:after+.
+    # @option rules [Boolean, nil] :symbol_after_without_space (false) Whether
+    #   a space between the amount and the money symbol should be inserted when
+    #   +:symbol_position+ is +:after+. The default is false (meaning space). Ignored
+    #   if +:symbol+ is false or +:symbol_position+ is not +:after+.
     #
     # @example
     #   # Default is to insert a space.
@@ -102,7 +102,7 @@ class Money
     #   # If set to true, will not insert a space.
     #   Money.new(100, "USD").format(:symbol_position => :after, :symbol_after_without_space => true) #=> "1.00$"
     #
-    # @option *rules [Boolean, String, nil] :decimal_mark (true) Whether the
+    # @option rules [Boolean, String, nil] :decimal_mark (true) Whether the
     #  currency should be separated by the specified character or '.'
     #
     # @example
@@ -113,7 +113,7 @@ class Money
     #   # to "." as decimal_mark.
     #   Money.new(100, "FOO").format #=> "$1.00"
     #
-    # @option *rules [Boolean, String, nil] :thousands_separator (true) Whether
+    # @option rules [Boolean, String, nil] :thousands_separator (true) Whether
     #  the currency should be delimited by the specified character or ','
     #
     # @example
@@ -129,14 +129,14 @@ class Money
     #   # default to "," as thousands_separator.
     #   Money.new(100000, "FOO").format #=> "$1,000.00"
     #
-    # @option *rules [Boolean] :html (false) Whether the currency should be
+    # @option rules [Boolean] :html (false) Whether the currency should be
     #  HTML-formatted. Only useful in combination with +:with_currency+.
     #
     # @example
-    #   s = Money.ca_dollar(570).format(:html => true, :with_currency => true)
-    #   s #=>  "$5.70 <span class=\"currency\">CAD</span>"
+    #   Money.ca_dollar(570).format(:html => true, :with_currency => true)
+    #   #=> "$5.70 <span class=\"currency\">CAD</span>"
     #
-    # @option *rules [Boolean] :sign_before_symbol (false) Whether the sign should be
+    # @option rules [Boolean] :sign_before_symbol (false) Whether the sign should be
     #  before the currency symbol.
     #
     # @example
@@ -145,7 +145,7 @@ class Money
     #   Money.new(-100, "GBP").format(:sign_before_symbol => false) #=> "£-1.00"
     #   Money.new(-100, "GBP").format                               #=> "£-1.00"
     #
-    # @option *rules [Boolean] :sign_positive (false) Whether positive numbers should be
+    # @option rules [Boolean] :sign_positive (false) Whether positive numbers should be
     #  signed, too.
     #
     # @example
@@ -156,7 +156,7 @@ class Money
     #   Money.new(100, "GBP").format(:sign_positive => false, :sign_before_symbol => false) #=> "£1.00"
     #   Money.new(100, "GBP").format                               #=> "£+1.00"
     #
-    # @option *rules [Boolean] :disambiguate (false) Prevents the result from being ambiguous
+    # @option rules [Boolean] :disambiguate (false) Prevents the result from being ambiguous
     #  due to equal symbols for different currencies. Uses the `disambiguate_symbol`.
     #
     # @example
@@ -165,21 +165,21 @@ class Money
     #   Money.new(10000, "USD").format(:disambiguate => true)    #=> "$100.00"
     #   Money.new(10000, "CAD").format(:disambiguate => true)    #=> "C$100.00"
     #
-    # @option *rules [Boolean] :html_wrap_symbol (false) Wraps the currency symbol
+    # @option rules [Boolean] :html_wrap_symbol (false) Wraps the currency symbol
     #  in a html <span> tag.
     #
     # @example
     #   Money.new(10000, "USD").format(:disambiguate => false)
     #   #=> "<span class=\"currency_symbol\">$100.00</span>
     #
-    # @option *rules [Symbol] :symbol_position (:before) `:before` if the currency
+    # @option rules [Symbol] :symbol_position (:before) `:before` if the currency
     #   symbol goes before the amount, `:after` if it goes after.
     #
     # @example
     #   Money.new(10000, "USD").format(:symbol_position => :before) #=> "$100.00"
     #   Money.new(10000, "USD").format(:symbol_position => :after)  #=> "100.00 $"
     #
-    # @option *rules [Boolean] :translate (true) `true` Checks for custom
+    # @option rules [Boolean] :translate (true) `true` Checks for custom
     #   symbol definitions using I18n.
     #
     # @example
@@ -195,9 +195,9 @@ class Money
     #   Money.new(89000, :btc).format(:drop_trailing_zeros => true) #=> B⃦0.00089
     #   Money.new(110, :usd).format(:drop_trailing_zeros => true)   #=> $1.1
     #
-    # Note that the default rules can be defined through +Money.default_formatting_rules+ hash.
+    # Note that the default rules can be defined through {Money.default_formatting_rules} hash.
     #
-    # @see +Money.default_formatting_rules+ for more information.
+    # @see Money.default_formatting_rules Money.default_formatting_rules for more information.
     
     def unformatted(*rules)
       rules = normalize_formatting_rules(rules)
@@ -218,6 +218,11 @@ class Money
       rules = default_formatting_rules.merge(rules)
       rules = localize_formatting_rules(rules)
       rules = translate_formatting_rules(rules) if rules[:translate]
+
+      thousands_separator = self.thousands_separator
+      decimal_mark = self.decimal_mark
+
+      escaped_decimal_mark = Regexp.escape(decimal_mark)
 
       if fractional == 0
         if rules[:display_free].respond_to?(:to_str)
@@ -258,9 +263,9 @@ class Money
 
       # Inspiration: https://github.com/rails/rails/blob/16214d1108c31174c94503caced3855b0f6bad95/activesupport/lib/active_support/number_helper/number_to_rounded_converter.rb#L72-L79
       if rules[:drop_trailing_zeros]
-        escaped_decimal_mark = Regexp.escape(decimal_mark)
         formatted = formatted.sub(/(#{escaped_decimal_mark})(\d*[1-9])?0+\z/, '\1\2').sub(/#{escaped_decimal_mark}\z/, '')
       end
+      has_decimal_value = !!(formatted =~ /#{escaped_decimal_mark}/)
 
       thousands_separator_value = thousands_separator
       # Determine thousands_separator
@@ -297,7 +302,7 @@ class Money
         formatted="#{sign_before}#{sign}#{formatted}"
       end
 
-      apply_decimal_mark_from_rules(formatted, rules)
+      apply_decimal_mark_from_rules(formatted, rules) if has_decimal_value
 
       if rules[:with_currency]
         formatted << " "
@@ -408,7 +413,11 @@ class Money
   def symbol_value_from(rules)
     if rules.has_key?(:symbol)
       if rules[:symbol] === true
-        symbol
+        if rules[:disambiguate] && currency.disambiguate_symbol
+          currency.disambiguate_symbol
+        else
+          symbol
+        end
       elsif rules[:symbol]
         rules[:symbol]
       else
@@ -416,7 +425,7 @@ class Money
       end
     elsif rules[:html]
       currency.html_entity == '' ? currency.symbol : currency.html_entity
-    elsif rules[:disambiguate] and currency.disambiguate_symbol
+    elsif rules[:disambiguate] && currency.disambiguate_symbol
       currency.disambiguate_symbol
     else
       symbol