+++ spec/money_spec.rb
@@ -191,7 +191,7 @@ describe Money do
     it "stores fractional as an integer regardless of what is passed into the constructor" do
       m = Money.new(100)
       expect(m.fractional).to eq 100
-      expect(m.fractional).to be_a(Fixnum)
+      expect(m.fractional).to be_a(Integer)
     end
 
     context "loading a serialized Money via YAML" do
@@ -349,9 +349,9 @@ YAML
       expect {money.round_to_nearest_cash_value}.to raise_error(Money::UndefinedSmallestDenomination)
     end
 
-    it "returns a Fixnum when infinite_precision is not set" do
+    it "returns a Integer when infinite_precision is not set" do
       money = Money.new(100, "USD")
-      expect(money.round_to_nearest_cash_value).to be_a Fixnum
+      expect(money.round_to_nearest_cash_value).to be_a Integer
     end
 
     it "returns a BigDecimal when infinite_precision is set", :infinite_precision do
@@ -408,12 +408,16 @@ YAML
   end
 
   describe "#currency_as_string=" do
-    it "sets the currency object using the provided string" do
+    it "sets the currency object using the provided string leaving cents intact" do
       money = Money.new(100_00, "USD")
+
       money.currency_as_string = "EUR"
       expect(money.currency).to eq Money::Currency.new("EUR")
+      expect(money.cents).to eq 100_00
+
       money.currency_as_string = "YEN"
       expect(money.currency).to eq Money::Currency.new("YEN")
+      expect(money.cents).to eq 100_00
     end
   end
 
@@ -582,6 +586,23 @@ YAML
       expect(moneys[0].cents).to eq 34
       expect(moneys[1].cents).to eq 33
       expect(moneys[2].cents).to eq 33
+    end
+
+    context "negative amount" do
+      it "does not lose pennies" do
+        moneys = Money.us_dollar(-100).allocate([0.333, 0.333, 0.333])
+
+        expect(moneys[0].cents).to eq -34
+        expect(moneys[1].cents).to eq -33
+        expect(moneys[2].cents).to eq -33
+      end
+
+      it "allocates the same way as positive amounts" do
+        ratios = [0.6667, 0.3333]
+
+        expect(Money.us_dollar(10_00).allocate(ratios).map(&:fractional)).to eq([6_67, 3_33])
+        expect(Money.us_dollar(-10_00).allocate(ratios).map(&:fractional)).to eq([-6_67, -3_33])
+      end
     end
 
     it "requires total to be less then 1" do